Fix missing alias for translations
I18n main lib provides two methods to translate copies. One is
`translate` and the other one is `t`. In this gem we are just overriding
`translate` but not `t`. So `t` is not adding global interpolation as
expected.

As you can see here, the original gem is doing that:

https://github.com/attilahorvath/i18n-globals/blob/master/lib/i18n/globals.rb#L118